Объединяет js библиотеки в vendor.js
