+++ .eslintignore
@@ -1,2 +1,1 @@
-source/js/slick.js
-source/js/jquery.min.js
+source/js/vendor.js
