+++ source/seat.html
@@ -287,8 +287,9 @@
       </div>
     </div>
   </footer>
-  <script src="js/jquery.min.js"></script>
-  <script src="js/slick.js"></script>
+  <!-- <script src="js/jquery.min.js"></script>
+  <script src="js/slick.js"></script> -->
+  <script src="js/vendor.js"></script>
   <script src="js/main.js"></script>
 </body>
 </html>
